% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverySimulator.R
\name{recoverySim}
\alias{recoverySim}
\title{Recovery simulator}
\usage{
recoverySim(
  simPar,
  cuPar,
  catchDat = NULL,
  srDat = NULL,
  variableCU = FALSE,
  makeSubDirs = TRUE,
  ricPars,
  larkPars = NULL,
  tamFRP = NULL,
  cuCustomCorrMat = NULL,
  erCorrMat = NULL,
  dirName,
  nTrials = 100,
  uniqueProd = TRUE,
  random = FALSE
)
}
\arguments{
\item{simPar}{Simulation parameters.}

\item{cuPar}{CU (Conservation Unit) parameters.}

\item{catchDat}{Historical catch data.}

\item{srDat}{Historical stock-recruitment data.}

\item{variableCU}{A logical to determine if CU name should be included in
output labels.}

\item{makeSubDirs}{Should sub-directories by made for outputs?}

\item{ricPars}{Ricker stock-recruitment parameters.}

\item{larkPars}{Larkin stock-recruitment parameters.}

\item{tamFRP}{Fisheries Reference Points for the Total Allowable Mortality
rule used for Fraser River Sockeye Salmon.}

\item{cuCustomCorrMat}{Custom correlation matrix for recruitment residuals.}

\item{erCorrMat}{Correlation matrix for exploitaion rates.}

\item{dirName}{Name of directory for outputs.}

\item{nTrials}{Number of Monte Carlo trials.}

\item{uniqueProd}{Should all CUs have unique productivities?}

\item{random}{Should all trials start with a random seed?}
}
\description{
Closed-loop simulation tool to assess management procedures and inform
Pacific salmon rebuilding strategies. Originally based on C. Holt's south
coast chum model (CSAS 2018). Simulation runs primed with observed SR data
(recDat input). The model includes data generation, variation in age
structure, survey design, and variable exploitation rules. OM uses Ricker
formulation \code{R=S(exp(a-bS))}. Additional details in changesToChumModel.md
in reports directory.
}
