% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESAsamSim.R
\docType{package}
\name{TESAsamSim}
\alias{TESAsamSim}
\title{TESAsamSim}
\description{
TESAsamSim is developed from samSim and is intended
to evaluate impacts of non-stationarity
for Pacific salmon. Adapted from samSim (by Cameron
Freshwater)
}
\author{
Carrie Holt <carrie.holt@dfo-mpo.gc.ca>
}
